+++ src/stdlib.rs
@@ -71,7 +71,7 @@ pub(crate) fn load_stdlib(context: &mut Rant)
 
   load_funcs!(
     // General functions
-    alt, call, either, len, get_type as "type", seed, nop, resolve, fork, unfork, try_ as "try",
+    alt, call, concat, either, len, get_type as "type", seed, nop, resolve, fork, unfork, try_ as "try",
 
     // Assertion functions
     assert as "assert", assert_eq as "assert-eq", assert_neq as "assert-neq",