+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New
 
 ### Changes
+  * Adjusted whitespace behavior around blocks to avoid ambiguities
 
 ### Fixes
   * Fixed incorrect whitespace print behavior around unflagged function calls