+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 4.0.0-alpha.16 (unreleased)
+## 4.0.0-alpha.16
 
 ### New
   * Added temporal argument spreading